Updated docs for select element
